dashboard/requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
plotly>=5.20.0
pyyaml>=6.0.1
python-difflib-stubs>=0.1.0
rich>=13.7.0
requests>=2.31.0
python-docx>=1.1.0
